Update dependencies and add integration test resources in build.gradle.kts
This commit updates the slf4j version used in build.gradle.kts and also adds srcDirs for integration-test resources in the sourceSets. Additionally, we've included the creation of a new log4j properties file, which will be used to manage logging during integration testing. These changes help to streamline the integration testing set up and improve the accuracy of logging.